commit 2

+++ application/controllers/client_dashboard.php
@@ -15,20 +15,34 @@ public function __construct()
 	}
 	public function index()
 	{
-		$clientid = $this->uri->segment(2); 
-		$userdata = array(
-            'client_id'  => $clientid,
-        );
-		$this->session->set_userdata($userdata);
-		$myuser_id = $this->session->userdata('client_id');
-		 $cookie= array(
-			'name'   => 'client',
-			'value'  => $myuser_id,
-			'expire' => '86500'
-		);
-		$this->input->set_cookie($cookie);
-		$this->db->truncate('fb');
-		$this->load->view('client_dashboard');
+		if($_POST['my_userid'] == "") {
+			$myuser_id = $this->session->userdata('client_id'); 
+			if($myuser_id == ''){
+					$myuser_id = $this->input->cookie('client',true);
+				}
+		}
+		//$clientid = $this->uri->segment(2); 
+		else {
+			$clientid = $_POST['my_userid']; 
+			$userdata = array(
+				'client_id'  => $clientid,
+			);
+			$this->session->set_userdata($userdata);
+			$myuser_id = $this->session->userdata('client_id');
+			 $cookie= array(
+				'name'   => 'client',
+				'value'  => $myuser_id,
+				'expire' => '86500'
+			);
+			$this->input->set_cookie($cookie);
+		}
+		if($myuser_id == "") {
+			redirect('login');
+		}
+		else{
+			$this->db->truncate('fb');
+			$this->load->view('client_dashboard');
+		}
 	}
 	
 }
